4791 - Store Refactor: Data->TableData->tableDataStatus and remove dataDep
