+++ src/client/pages/Section/DataTable/hooks/useData/useData.ts
@@ -4,7 +4,7 @@ import { Objects } from 'utils/objects'
 
 import { RecordAssessmentData, RecordAssessmentDatas } from 'meta/data'
 
-import { useIsSomeTableDataFetching } from 'client/store/data'
+import { useIsSomeTableDataFetching } from 'client/store/data/tableData/status/hooks/tableDataStatus'
 import { useAssessment } from 'client/store/meta/hooks/assessments'
 import { useCycle } from 'client/store/meta/hooks/cycles'
 import { useCountryIso } from 'client/hooks'
