+++ src/client/pages/Section/Title/ForestCharacteristics/ForestCharacteristics.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'react-i18next'
 import { Labels } from 'meta/assessment/labels'
 
 import { useAssessmentCountry } from 'client/store/area/hooks/country'
-import { useOriginalDataPointYears } from 'client/store/data'
+import { useOriginalDataPointYears } from 'client/store/data/tableData/nodeValues/hooks/data'
 import { useCycle } from 'client/store/meta/hooks/cycles'
 import { useIsPrintRoute } from 'client/hooks/useIsRoute'
 import OriginalDataPointsPrint from 'client/pages/Print/OriginalDataPointsPrint'
