--- src/client/store/data/hooks/useHasOriginalDataPointData.ts
@@ -1,3 +1,0 @@
-import { useOriginalDataPointData } from 'client/store/data/hooks/useOriginalDataPointData'
-
-export const useHasOriginalDataPointData = (): boolean => Object.keys(useOriginalDataPointData()).length > 0
