+++ src/client/store/data/tableData/nodeValues/hooks/data.ts
@@ -1,19 +1,31 @@
-import { useMemo } from 'react'
+import { useEffect, useMemo, useState } from 'react'
 
 import { Objects } from 'utils/objects'
 
+import { CountryIso } from 'meta/area'
 import { TableName, TableNames } from 'meta/assessment/table'
-import { RecordAssessmentData, RecordAssessmentDatas, RecordTableData } from 'meta/data'
+import { TableSection } from 'meta/assessment/tableSection'
+import { RecordAssessmentData, RecordAssessmentDatas, RecordColumnData, RecordTableData } from 'meta/data'
 
 import { useAssessmentCountry } from 'client/store/area/hooks/country'
-import { useOriginalDataPointData } from 'client/store/data/hooks/useOriginalDataPointData'
 import { NodeValuesSelectors } from 'client/store/data/tableData/nodeValues/selectors'
 import { useAppSelector } from 'client/store/hooks'
 import { useAssessment } from 'client/store/meta/hooks/assessments'
 import { useCycle } from 'client/store/meta/hooks/cycles'
+import { useCountryRouteParams } from 'client/hooks/useRouteParams'
 
 export const useRecordAssessmentData = (): RecordAssessmentData =>
   useAppSelector(NodeValuesSelectors.getRecordAssessmentData)
+
+export const useOriginalDataPointData = (): RecordColumnData => {
+  const tableName = TableNames.originalDataPointValue
+  const { assessmentName, countryIso, cycleName } = useCountryRouteParams<CountryIso>()
+
+  return useAppSelector((state) => {
+    const data = NodeValuesSelectors.getRecordAssessmentData(state)
+    return RecordAssessmentDatas.getTableData({ assessmentName, cycleName, data, countryIso, tableName })
+  })
+}
 
 export const useRecordAssessmentDataWithOdp = (): RecordAssessmentData => {
   const assessment = useAssessment()
@@ -60,4 +72,66 @@ export const useRecordAssessmentDataWithOdp = (): RecordAssessmentData => {
       },
     }
   }, [assessmentName, country.props.forestCharacteristics.useOriginalDataPoint, countryIso, cycleName, data, odpData])
+}
+
+// export const useIsOdpTableDataFetched = (): boolean => {
+//   const { assessmentName, countryIso, cycleName } = useCountryRouteParams<CountryIso>()
+//   const data = useRecordAssessmentData()
+//
+//   return useMemo<boolean>(() => {
+//     return Object.hasOwn(
+//       RecordAssessmentDatas.getCountryData({ assessmentName, cycleName, countryIso, data }),
+//       TableNames.originalDataPointValue
+//     )
+//   }, [assessmentName, countryIso, cycleName, data])
+// }
+
+export const useIsSectionDataEmpty = (tableSections: Array<TableSection>) => {
+  const { assessmentName, countryIso, cycleName } = useCountryRouteParams<CountryIso>()
+
+  const data = useRecordAssessmentData()
+
+  const [sectionDataEmpty, setSectionDataEmpty] = useState(false)
+  const sectionTableNames = useMemo(
+    () => tableSections.flatMap((ts) => ts.tables.flatMap((t) => t.props.name)),
+    [tableSections]
+  )
+
+  const dataLoaded = useMemo(
+    () => Boolean(data?.[assessmentName]?.[cycleName]?.[countryIso]),
+    [assessmentName, countryIso, cycleName, data]
+  )
+
+  const allTablesEmpty =
+    dataLoaded &&
+    sectionTableNames.every((tableName) =>
+      RecordAssessmentDatas.isTableDataEmpty({
+        assessmentName,
+        cycleName,
+        data,
+        tableName,
+        countryIso,
+      })
+    )
+
+  useEffect(() => {
+    if (dataLoaded) {
+      setSectionDataEmpty(allTablesEmpty)
+    }
+  }, [allTablesEmpty, dataLoaded])
+
+  if (!dataLoaded) return false
+
+  return sectionDataEmpty
+}
+
+export const useOriginalDataPointYears = (): Array<{ year: string; id: number }> => {
+  const odpData = useOriginalDataPointData()
+
+  if (Objects.isEmpty(odpData)) return null
+
+  return Object.entries(odpData).map(([year, data]) => ({
+    year,
+    id: data.totalLandArea.odpId,
+  }))
 }
