+++ src/client/store/data/tableData/status/slice/index.ts
@@ -1,0 +1,13 @@
+import { ActionReducerMapBuilder, createSlice } from '@reduxjs/toolkit'
+
+import { getTableDataReducer } from 'client/store/data/tableData/status/slice/extraReducers/getTableDataReducer'
+import { TableDataStatusState } from 'client/store/data/tableData/status/state'
+
+export const TableDataStatusSlice = createSlice({
+  name: 'tableDataStatus',
+  initialState: {},
+  reducers: {},
+  extraReducers: (builder: ActionReducerMapBuilder<TableDataStatusState>) => {
+    getTableDataReducer(builder)
+  },
+})
